Using AbstractNetworkParams
* using `InetAddress.getLocalHost().getHostName()` for getHostName
* put resolv.conf parsing code to its getDnsServers
